Values of input form are changing and computing
