+++ packages/bonde-admin-canary/src/scenes/Auth/scenes/ResetPassword/ResetPasswordForm/ResetPasswordForm.js
@@ -1,20 +1,39 @@
 import React from 'react'
+import PropTypes from 'prop-types'
+
+import { required, min } from 'services/validations'
+import { AuthAPI } from 'services/auth'
+import resetPassword from './resetPassword.graphql'
+
 import { Flexbox2 as Flexbox, Title } from 'bonde-styleguide'
 import { Field, MutationForm, SubmitButton } from 'components/Forms'
 import { ButtonLink } from 'components/Link'
-import { required, min } from 'services/validations'
+import { notify } from 'components/Notification'
 import { PasswordField } from '../../components'
-import resetPassword from './resetPassword.graphql'
-import PropTypes from 'prop-types'
 
-const ResetPasswordForm = ({ t, token, handleSuccess }) => (
+const formName = 'ResetPasswordForm'
+
+const ResetPasswordForm = ({ t, token }) => (
   <Flexbox vertical>
     <Title.H2 margin={{ bottom: 18 }}>{t('resetPassword.form.title')}</Title.H2>
     <Title.H4 margin={{ bottom: 30 }}>{t('resetPassword.form.subtitle')}</Title.H4>
     <MutationForm
       mutation={resetPassword}
       variables={{ token }}
-      onSuccess={handleSuccess}
+      onSuccess={({ data }) => {
+        const { changePasswordField } = data.resetPasswordChangePassword
+        const user = { name: changePasswordField.userFirstName }
+
+        AuthAPI.login({
+          jwtToken: changePasswordField.token
+        })
+          .then(() => {
+            notify(t('resetPassword.success', { user }))
+            // should redirect with window to rehydrate session
+            window.location.href = '/'
+          })
+      }}
+      formId={formName}
     >
       <Field
         name='password'
@@ -28,7 +47,7 @@ const ResetPasswordForm = ({ t, token, handleSuccess }) => (
       />
       <Flexbox padding={{ top: 25 }} horizontal spacing='between'>
         <ButtonLink to='/auth/login'>{t('resetPassword.form.cancel')}</ButtonLink>
-        <SubmitButton>{t('resetPassword.form.submit')}</SubmitButton>
+        <SubmitButton title={t('resetPassword.form.submit')} formId={formName}>{t('resetPassword.form.submit')}</SubmitButton>
       </Flexbox>
     </MutationForm>
   </Flexbox>
